Fix singleton destruction order (#1748)
* Refs 10533 Fix singleton destructin order

These singletons are used on the destruction of certain classes.
In order to avoid these singletons to be already destructed,
these classes keep a shared_pointer to an instance to the singleton.
This ensures that the singleton is destructed AFTER the class instance.


* Refs 10533 Uncrustify


* Refs 10533 Resolve warnings


* Refs 10533 Avoid copy on instance construction


* Refs 10533 use create_entities on the new test

+++ test/unittest/dds/subscriber/DataReaderTests.cpp
@@ -77,6 +77,11 @@ public:
 
     void TearDown() override
     {
+        if (!destroy_entities_)
+        {
+            return;
+        }
+
         if (data_writer_)
         {
             ASSERT_EQ(publisher_->delete_datawriter(data_writer_), ReturnCode_t::RETCODE_OK);
@@ -528,6 +533,7 @@ protected:
     DataReader* data_reader_ = nullptr;
     DataWriter* data_writer_ = nullptr;
     TypeSupport type_;
+    bool destroy_entities_ = true;
 
     InstanceHandle_t handle_ok_ = HANDLE_NIL;
     InstanceHandle_t handle_wrong_ = HANDLE_NIL;
@@ -1270,6 +1276,12 @@ TEST_F(DataReaderTests, read_unread)
             EXPECT_EQ(ok_code, data_reader_->return_loan(data_seq[i], info_seq[i]));
         }
     }
+}
+
+TEST_F(DataReaderTests, TerminateWithoutDestroyingReader)
+{
+    destroy_entities_ = false;
+    create_entities();
 }
 
 void set_listener_test (
